backend/requirements-test.txt
# Testing dependencies for Market Sector Sentiment Analysis Tool

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# HTTP testing
httpx>=0.24.0
pytest-httpx>=0.24.0

# Performance testing
pytest-benchmark>=4.0.0

# Test data generation
factory-boy>=3.3.0
faker>=19.0.0

# Property-based testing
hypothesis>=6.75.0

# Coverage reporting
coverage>=7.2.0

# Test reporting
pytest-html>=3.2.0
pytest-xdist>=3.3.0

# Mocking and stubbing
responses>=0.23.0
freezegun>=1.2.0

# Database testing
pytest-postgresql>=4.1.0
pytest-redis>=2.0.0

# Environment management
python-dotenv>=1.0.0 